Move existing button to deprecated folder
