+++ packages/components/Button/src/deprecated/Button.settings.win32.ts
