+++ packages/components/Button/src/deprecated/StealthButton/StealthButton.settings.ts
