+++ packages/components/Button/src/deprecated/StealthButton/StealthButton.ts
