+++ packages/components/Button/src/deprecated/StealthButton/index.ts
