+++ packages/components/Button/src/index.ts
@@ -1,4 +1,4 @@
-export * from './Button.types';
-export * from './Button';
-export * from './PrimaryButton';
-export * from './StealthButton';
+export * from './deprecated/Button.types';
+export * from './deprecated/Button';
+export * from './deprecated/PrimaryButton';
+export * from './deprecated/StealthButton';
